+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/ResourceListTest.java
@@ -27,8 +27,13 @@ import io.fabric8.kubernetes.api.model.PodBuilder;
 import io.fabric8.kubernetes.api.model.PodListBuilder;
 import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.ServiceBuilder;
+import io.fabric8.kubernetes.api.model.Status;
+import io.fabric8.kubernetes.api.model.StatusBuilder;
+import io.fabric8.kubernetes.api.model.WatchEvent;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientException;
+import io.fabric8.kubernetes.client.KubernetesClientTimeoutException;
+import io.fabric8.kubernetes.client.dsl.ListVisitFromServerGetDeleteRecreateWaitApplicable;
 import io.fabric8.kubernetes.client.dsl.NamespaceListVisitFromServerGetDeleteRecreateWaitApplicable;
 import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
 
@@ -38,10 +43,18 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.migrationsupport.rules.EnableRuleMigrationSupport;
 
-import java.net.HttpURLConnection;
+import java.time.Duration;
 import java.util.Collections;
 import java.util.List;
+import java.util.function.Predicate;
 
+import static java.net.HttpURLConnection.HTTP_CONFLICT;
+import static java.net.HttpURLConnection.HTTP_CREATED;
+import static java.net.HttpURLConnection.HTTP_GONE;
+import static java.net.HttpURLConnection.HTTP_OK;
+import static java.net.HttpURLConnection.HTTP_UNAVAILABLE;
+import static java.time.Duration.ofSeconds;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -77,7 +90,7 @@ public class ResourceListTest {
   void testCreateOrReplace() {
     Pod pod1 = new PodBuilder().withNewMetadata().withName("pod1").withNamespace("test").and().build();
 
-    server.expect().post().withPath("/api/v1/namespaces/test/pods").andReturn(HttpURLConnection.HTTP_CREATED, pod1).once();
+    server.expect().post().withPath("/api/v1/namespaces/test/pods").andReturn(HTTP_CREATED, pod1).once();
 
     List<HasMetadata> response = client.resourceList(new PodListBuilder().addToItems(pod1).build()).createOrReplace();
     assertTrue(response.contains(pod1));
@@ -87,7 +100,7 @@ public class ResourceListTest {
   void testCreateOrReplaceFailedCreate() {
     // Given
     Pod pod1 = new PodBuilder().withNewMetadata().withName("pod1").withNamespace("test").and().build();
-    server.expect().post().withPath("/api/v1/namespaces/test/pods").andReturn(HttpURLConnection.HTTP_UNAVAILABLE, pod1).once();
+    server.expect().post().withPath("/api/v1/namespaces/test/pods").andReturn(HTTP_UNAVAILABLE, pod1).once();
     NamespaceListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata, Boolean> listOp = client.resourceList(new PodListBuilder().addToItems(pod1).build());
 
     // When
@@ -98,7 +111,7 @@ public class ResourceListTest {
   void testCreateWithExplicitNamespace() {
     Pod pod1 = new PodBuilder().withNewMetadata().withName("pod1").withNamespace("test").and().build();
 
-    server.expect().post().withPath("/api/v1/namespaces/ns1/pods").andReturn(HttpURLConnection.HTTP_CREATED, pod1).once();
+    server.expect().post().withPath("/api/v1/namespaces/ns1/pods").andReturn(HTTP_CREATED, pod1).once();
 
     List<HasMetadata> response = client.resourceList(new PodListBuilder().addToItems(pod1).build()).inNamespace("ns1").apply();
     assertTrue(response.contains(pod1));
@@ -111,9 +124,9 @@ public class ResourceListTest {
     Pod pod3 = new PodBuilder().withNewMetadata().withName("pod3").withNamespace("any").and().build();
 
 
-    server.expect().withPath("/api/v1/namespaces/test/pods/pod1").andReturn(HttpURLConnection.HTTP_OK, pod1).times(2);
-    server.expect().withPath("/api/v1/namespaces/ns1/pods/pod2").andReturn(HttpURLConnection.HTTP_OK, pod2).times(2);
-    server.expect().withPath("/api/v1/namespaces/any/pods/pod3").andReturn(HttpURLConnection.HTTP_OK, pod3).times(1);
+    server.expect().withPath("/api/v1/namespaces/test/pods/pod1").andReturn(HTTP_OK, pod1).times(2);
+    server.expect().withPath("/api/v1/namespaces/ns1/pods/pod2").andReturn(HTTP_OK, pod2).times(2);
+    server.expect().withPath("/api/v1/namespaces/any/pods/pod3").andReturn(HTTP_OK, pod3).times(1);
 
     //First time all items should be deleted.
     Boolean deleted = client.resourceList(new PodListBuilder().withItems(pod1, pod2, pod3).build()).delete();
@@ -126,12 +139,12 @@ public class ResourceListTest {
 
   @Test
   void testCreateOrReplaceWithoutDeleteExisting() throws Exception {
-    server.expect().post().withPath("/api/v1/namespaces/ns1/services").andReturn(HttpURLConnection.HTTP_CONFLICT, service).once();
-    server.expect().post().withPath("/api/v1/namespaces/ns1/configmaps").andReturn(HttpURLConnection.HTTP_CONFLICT, configMap).once();
-    server.expect().get().withPath("/api/v1/namespaces/ns1/services/my-service").andReturn(HttpURLConnection.HTTP_OK , service).once();
-    server.expect().get().withPath("/api/v1/namespaces/ns1/configmaps/my-configmap").andReturn(HttpURLConnection.HTTP_OK, configMap).once();
-    server.expect().put().withPath("/api/v1/namespaces/ns1/services/my-service").andReturn(HttpURLConnection.HTTP_OK, updatedService).once();
-    server.expect().put().withPath("/api/v1/namespaces/ns1/configmaps/my-configmap").andReturn(HttpURLConnection.HTTP_OK, updatedConfigMap).once();
+    server.expect().post().withPath("/api/v1/namespaces/ns1/services").andReturn(HTTP_CONFLICT, service).once();
+    server.expect().post().withPath("/api/v1/namespaces/ns1/configmaps").andReturn(HTTP_CONFLICT, configMap).once();
+    server.expect().get().withPath("/api/v1/namespaces/ns1/services/my-service").andReturn(HTTP_OK , service).once();
+    server.expect().get().withPath("/api/v1/namespaces/ns1/configmaps/my-configmap").andReturn(HTTP_OK, configMap).once();
+    server.expect().put().withPath("/api/v1/namespaces/ns1/services/my-service").andReturn(HTTP_OK, updatedService).once();
+    server.expect().put().withPath("/api/v1/namespaces/ns1/configmaps/my-configmap").andReturn(HTTP_OK, updatedConfigMap).once();
 
     client.resourceList(resourcesToUpdate).inNamespace("ns1").createOrReplace();
 
@@ -143,12 +156,12 @@ public class ResourceListTest {
 
   @Test
   void testCreateOrReplaceWithDeleteExisting() throws Exception {
-    server.expect().post().withPath("/api/v1/namespaces/ns1/services").andReturn(HttpURLConnection.HTTP_CONFLICT, service).once();
-    server.expect().post().withPath("/api/v1/namespaces/ns1/configmaps").andReturn(HttpURLConnection.HTTP_CONFLICT, configMap).once();
-    server.expect().delete().withPath("/api/v1/namespaces/ns1/services/my-service").andReturn(HttpURLConnection.HTTP_OK , service).once();
-    server.expect().delete().withPath("/api/v1/namespaces/ns1/configmaps/my-configmap").andReturn(HttpURLConnection.HTTP_OK, configMap).once();
-    server.expect().post().withPath("/api/v1/namespaces/ns1/services").andReturn(HttpURLConnection.HTTP_OK, updatedService).once();
-    server.expect().post().withPath("/api/v1/namespaces/ns1/configmaps").andReturn(HttpURLConnection.HTTP_OK, updatedConfigMap).once();
+    server.expect().post().withPath("/api/v1/namespaces/ns1/services").andReturn(HTTP_CONFLICT, service).once();
+    server.expect().post().withPath("/api/v1/namespaces/ns1/configmaps").andReturn(HTTP_CONFLICT, configMap).once();
+    server.expect().delete().withPath("/api/v1/namespaces/ns1/services/my-service").andReturn(HTTP_OK , service).once();
+    server.expect().delete().withPath("/api/v1/namespaces/ns1/configmaps/my-configmap").andReturn(HTTP_OK, configMap).once();
+    server.expect().post().withPath("/api/v1/namespaces/ns1/services").andReturn(HTTP_OK, updatedService).once();
+    server.expect().post().withPath("/api/v1/namespaces/ns1/configmaps").andReturn(HTTP_OK, updatedConfigMap).once();
 
     client.resourceList(resourcesToUpdate).inNamespace("ns1").deletingExisting().createOrReplace();
 
@@ -156,6 +169,167 @@ public class ResourceListTest {
     RecordedRequest request = server.getLastRequest();
     assertEquals("/api/v1/namespaces/ns1/configmaps", request.getPath());
     assertEquals("POST", request.getMethod());
+  }
+
+  @Test
+  void testSuccessfulWaitUntilCondition() throws InterruptedException {
+    Pod pod1 = new PodBuilder().withNewMetadata()
+      .withName("pod1")
+      .withResourceVersion("1")
+      .withNamespace("ns1").and().build();
+    Pod noReady1 = createReadyFrom(pod1, "False");
+    Pod ready1 = createReadyFrom(pod1, "True");
+
+    Pod pod2 = new PodBuilder().withNewMetadata()
+      .withName("pod2")
+      .withResourceVersion("1")
+      .withNamespace("ns1").and().build();
+    Pod noReady2 = createReadyFrom(pod2, "False");
+    Pod ready2 = createReadyFrom(pod2, "True");
+
+    Predicate<HasMetadata> isReady = p -> "Pod".equals(p.getKind()) && ((Pod) p).getStatus().getConditions().stream()
+      .anyMatch(c -> "True".equals(c.getStatus()));
+
+    // The pods are never ready if you request them directly.
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/pod1").andReturn(200, noReady1).once();
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/pod2").andReturn(200, noReady2).once();
+
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?fieldSelector=metadata.name%3Dpod1&watch=true").andUpgradeToWebSocket()
+      .open()
+      .waitFor(500).andEmit(new WatchEvent(ready1, "MODIFIED"))
+      .done()
+      .once();
+
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?fieldSelector=metadata.name%3Dpod2&watch=true").andUpgradeToWebSocket()
+      .open()
+      .waitFor(500).andEmit(new WatchEvent(ready2, "MODIFIED"))
+      .done()
+      .once();
+
+    try (KubernetesClient client = server.getClient()) {
+      KubernetesList list = new KubernetesListBuilder().withItems(pod1, pod2).build();
+      List<HasMetadata> results = client.resourceList(list).inNamespace("ns1")
+        .waitUntilCondition(isReady, null, ofSeconds(5));
+      assertThat(results)
+        .containsExactlyInAnyOrder(ready1, ready2);
+    }
+  }
+
+  @Test
+  void testPartialSuccessfulWaitUntilCondition() {
+    Pod pod1 = new PodBuilder().withNewMetadata()
+      .withName("pod1")
+      .withResourceVersion("1")
+      .withNamespace("ns1").and().build();
+    Pod noReady1 = createReadyFrom(pod1, "False");
+
+    Pod pod2 = new PodBuilder().withNewMetadata()
+      .withName("pod2")
+      .withResourceVersion("1")
+      .withNamespace("ns1").and().build();
+    Pod noReady2 = createReadyFrom(pod2, "False");
+    Pod ready2 = createReadyFrom(pod2, "True");
+
+    Predicate<HasMetadata> isReady = p -> "Pod".equals(p.getKind()) && ((Pod) p).getStatus().getConditions().stream()
+      .anyMatch(c -> "True".equals(c.getStatus()));
+
+    // The pods are never ready if you request them directly.
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/pod1").andReturn(200, noReady1).once();
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/pod2").andReturn(200, noReady2).once();
+
+    Status gone = new StatusBuilder()
+      .withCode(HTTP_GONE)
+      .build();
+
+    // This pod has a non-retryable error.
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?fieldSelector=metadata.name%3Dpod1&watch=true")
+      .andUpgradeToWebSocket()
+      .open()
+      .waitFor(500).andEmit(new WatchEvent(gone, "ERROR"))
+      .done()
+      .once();
+
+    // This pod succeeds.
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?fieldSelector=metadata.name%3Dpod2&watch=true")
+      .andUpgradeToWebSocket()
+      .open()
+      .waitFor(500).andEmit(new WatchEvent(ready2, "MODIFIED"))
+      .done()
+      .once();
+
+    try (KubernetesClient client = server.getClient()) {
+      KubernetesList list = new KubernetesListBuilder().withItems(pod1, pod2).build();
+      final ListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata, Boolean> ops = client.resourceList(list).inNamespace("ns1");
+      final Duration timeout = ofSeconds(5);
+      KubernetesClientTimeoutException ex = assertThrows(KubernetesClientTimeoutException.class, () ->
+        ops.waitUntilCondition(isReady, null, timeout)
+      );
+      assertThat(ex.getResourcesNotReady())
+        .containsExactly(pod1);
+    }
+  }
+
+  @Test
+  void testAllFailedWaitUntilCondition() {
+    Pod pod1 = new PodBuilder().withNewMetadata()
+      .withName("pod1")
+      .withResourceVersion("1")
+      .withNamespace("ns1").and().build();
+    Pod noReady1 = createReadyFrom(pod1, "False");
+
+    Pod pod2 = new PodBuilder().withNewMetadata()
+      .withName("pod2")
+      .withResourceVersion("1")
+      .withNamespace("ns1").and().build();
+    Pod noReady2 = createReadyFrom(pod2, "False");
+
+    Predicate<HasMetadata> isReady = p -> "Pod".equals(p.getKind()) && ((Pod) p).getStatus().getConditions().stream()
+      .anyMatch(c -> "True".equals(c.getStatus()));
+
+    // The pods are never ready if you request them directly.
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/pod1").andReturn(200, noReady1).once();
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/pod2").andReturn(200, noReady2).once();
+
+    Status gone = new StatusBuilder()
+      .withCode(HTTP_GONE)
+      .build();
+
+    // Both pods have a non-retryable error.
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?fieldSelector=metadata.name%3Dpod1&watch=true")
+      .andUpgradeToWebSocket()
+      .open()
+      .waitFor(500).andEmit(new WatchEvent(gone, "ERROR"))
+      .done()
+      .once();
+
+    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?fieldSelector=metadata.name%3Dpod2&watch=true")
+      .andUpgradeToWebSocket()
+      .open()
+      .waitFor(500).andEmit(new WatchEvent(gone, "ERROR"))
+      .done()
+      .once();
+
+    try (KubernetesClient client = server.getClient()) {
+      KubernetesList list = new KubernetesListBuilder().withItems(pod1, pod2).build();
+      final ListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata, Boolean> ops = client.resourceList(list).inNamespace("ns1");
+      final Duration timeout = ofSeconds(5);
+      KubernetesClientTimeoutException ex = assertThrows(KubernetesClientTimeoutException.class, () ->
+        ops.waitUntilCondition(isReady, null, timeout)
+      );
+      assertThat(ex.getResourcesNotReady())
+        .containsExactlyInAnyOrder(pod1, pod2);
+    }
+  }
+
+  private static Pod createReadyFrom(Pod pod, String status) {
+    return new PodBuilder(pod)
+      .withNewStatus()
+        .addNewCondition()
+          .withType("Ready")
+          .withStatus(status)
+        .endCondition()
+      .endStatus()
+      .build();
   }
 
   private static ServiceBuilder mockService() {