+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 #### Improvements
 * Fix #2331: Fixed documentation for namespaced informer for all custom types implementing `Namespaced` interface
 * Fix #2406: Add documentation for serializing resources to YAML
+* Fix #2414: Add waitUntilCondition() operation that accepts a new resource version. This enables recovery from HTTP 410 GONE errors.
 
 #### Dependency Upgrade
 * Fix #2360: bump mockito-core from 3.4.0 to 3.4.2